Make "=" work as in most other languages, and add ":=" for previous behavior.
This implemented RFC on making "<-" operator optional:
http://forum.squirrel-lang.org/mainsite/forums/default.aspx?g=posts&t=3113
"=" is made to behave as "<-" operator when it applies to table slots
(it still behaves as before for local variables, in other words, "=" can
now be used consistently for all of: global vars, member vars, local vars,
dictionary/mapping/table key assignment). Old behavior of "=" is preserved
in newly introduced ":=" operator (i.e. it won't create a new slot if
applied to a table-backed symtab). Note that "<-" operator stays the same,
user just no longer bound to use it with new "=".

Summing up, old code:

a <- 1
a = 2

Can now be written as:

a = 1
a = 2

Alternatively, too fully preserve old semantics, it can be written as:

a <- 1
a := 2