Add elaborated makefile to build Squirrel.
This is better makefile for Squirrel, part of General-Purpose Squirrel
project. The trick is that GNU Make looks for "makefile", then "Makefile",
so this file overrides upstream Makefile (which is not ideal and laeft in
place only to make it easier merge with upstream). This however means
that there may be problems with case-insensitive file systems, so don't
use such, use FSes with proper POSIX semantics.

This file support specifying compile and link - type options in one places,
easy overriding of them, and supporting build of different versions of
targets with different options - in different dirs, using VPATH feature.